backend/requirements-security.txt
# Authentication and JWT handling
python-jose[cryptography]==3.3.0  # JWT token creation and validation

# Password hashing
passlib[bcrypt]==1.7.4  # Secure password hashing with bcrypt

# Rate limiting - core only, Redis integration separate
slowapi==0.1.8  # Basic rate limiting functionality

# Security headers - minimal install
secure==0.3.0  # CSP, HSTS, and security headers

# Redis backend for distributed rate limiting
redis==5.0.1  # Required for distributed rate limiting
aioredis==2.0.1  # Async Redis client for better performance

# Monitoring and metrics
prometheus-client==0.17.1  # Prometheus metrics export